Autorizacija admina pomaknuta u Authorizer klasu
