+++ src/main/java/foi/air/szokpt/accountmng/services/UserService.java
@@ -2,12 +2,10 @@ package foi.air.szokpt.accountmng.services;
 
 import foi.air.szokpt.accountmng.entitites.User;
 import foi.air.szokpt.accountmng.entitites.UserRole;
-import foi.air.szokpt.accountmng.exceptions.AuthorizationException;
 import foi.air.szokpt.accountmng.exceptions.NotFoundException;
 import foi.air.szokpt.accountmng.exceptions.ValidationException;
 import foi.air.szokpt.accountmng.repositories.RoleRepository;
 import foi.air.szokpt.accountmng.repositories.UserRepository;
-import foi.air.szokpt.accountmng.util.JwtUtil;
 import foi.air.szokpt.accountmng.util.hashing.Hasher;
 import foi.air.szokpt.accountmng.util.validation.Validator;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -22,22 +20,19 @@ public class UserService {
     private final Validator<User> updateUserValidator;
     private final Validator<User> registerUserValidator;
     private final Hasher hasher;
-    private final JwtUtil jwtUtil;
 
     public UserService(UserRepository userRepository, RoleRepository roleRepository,
                        @Qualifier("registerUserValidator") Validator<User> registerUserValidator,
                        @Qualifier("updateUserValidator") Validator<User> updateUserValidator,
-                       Hasher hasher, JwtUtil jwtUtil) {
+                       Hasher hasher) {
         this.userRepository = userRepository;
         this.roleRepository = roleRepository;
         this.updateUserValidator = updateUserValidator;
         this.registerUserValidator = registerUserValidator;
         this.hasher = hasher;
-        this.jwtUtil = jwtUtil;
     }
 
-    public void registerUser(String authorizationHeader, User user) {
-        authorizeAdmin(authorizationHeader);
+    public void registerUser(User user) {
         registerUserValidator.validateData(user);
         assignRoleToUser(user, user.getRole().getName());
         user.setEmail(user.getEmail().toLowerCase());
@@ -45,8 +40,7 @@ public class UserService {
         userRepository.save(user);
     }
 
-    public void updateUser(int id, User newUserData, String authorizationHeader) {
-        authorizeAdmin(authorizationHeader);
+    public void updateUser(int id, User newUserData) {
         Optional<User> optionalExistingUser = userRepository.findById(id);
         if (optionalExistingUser.isPresent()) {
             newUserData.setId(id);
@@ -66,23 +60,6 @@ public class UserService {
         assignRoleToUser(existingUser, newUserData.getRole().getName());
         existingUser.setPassword(hasher.hashText(newUserData.getPassword()));
         userRepository.save(existingUser);
-    }
-
-    private void authorizeAdmin(String authorizationHeader) {
-        String token = jwtUtil.extractToken(authorizationHeader);
-        verifyToken(token);
-        checkAdminRole(token);
-    }
-
-    private void verifyToken(String token) {
-        jwtUtil.verifyToken(token);
-    }
-
-    private void checkAdminRole(String token) {
-        String userRole = jwtUtil.getRoleName((token));
-        if (!userRole.equals("admin")) {
-            throw new AuthorizationException("User role is not authorized for this action");
-        }
     }
 
     private void assignRoleToUser(User user, String roleName) {
